Move deprecated decorator to utils
